ソースコード/requirements.txt
# Django
Django~=4.2
django-environ~=0.11.0
djangorestframework~=3.16.0

# camelcase
djangorestframework-camel-case~=1.4.0

# CORS
django-cors-headers~=4.3.0

# server
uwsgi~=2.0.26

# auth
django-allauth~=0.57.0
dj-rest-auth~=5.0.0

# DB
psycopg2~=2.9.9

# linting and formatting
flake8
black
isort

# s3
Pillow~=10.4.0
django-storages~=1.14.0
boto3~=1.35.0

# admin
django-import-export~=4.1.0
django-reverse-admin~=2.9.6
